requirements.txt
graphene==2.1.8
starlette==0.14.2
uvicorn==0.13.3
pyfunctional==1.4.3
